CompanyX
Company X generally processes orders based on a first come, first serve basis.  Sometimes a client will ask for rush processing and will be awarded a number based on the urgency. If they pay extra they will receive a 3 priority, if they have a bulk order they will receive a 2 priority, and if they know someone in the company and ask for a favor they will receive the highest priority, a 1.  Designed a system that accepts orders from clients and processes these orders based on no priority or somewhere on the priority scale. Showed how the system will work for random orders with a random processing time.